List3: Finish

+++ list3/zad1/src/main/java/dijkstra/Dijkstra.java
@@ -10,16 +10,16 @@ public class Dijkstra {
 
     private DirectedWeightedGraph graph;
     private int begin;
-    private int[] d;
+    private double[] d;
     private int[] pi;
-    private PriorityQueue<Integer> Q;
+    private PriorityQueue<Double> Q;
 
     public Dijkstra(DirectedWeightedGraph graph) {
         this.graph = graph;
     }
 
     private void initializeSingleSource() {
-        d = new int[graph.verticesCount()];
+        d = new double[graph.verticesCount()];
         pi = new int[graph.verticesCount()];
         for (int i = 0; i < graph.verticesCount(); i++) {
             d[i] = Integer.MAX_VALUE;
@@ -36,12 +36,12 @@ public class Dijkstra {
         }
     }
 
-    public int[] shortestPath(int begin) {
+    public double[] shortestPath(int begin) {
 
         this.begin = begin;
         initializeSingleSource();
 
-        Q = new PriorityQueue<>(graph.verticesCount(), Integer.MAX_VALUE);
+        Q = new PriorityQueue<>(graph.verticesCount(), Double.MAX_VALUE);
         for (int i = 0; i < graph.verticesCount(); i++) {
             Q.insert(i, d[i]);
         }
