+++ list3/zad1/src/main/java/dijkstra/DirectedWeightedEdge.kt
@@ -2,7 +2,7 @@ package dijkstra
 
 import graph.WeightedEdge
 
-data class DirectedWeightedEdge(val begin: Int, val end: Int, val weight: Int) : WeightedEdge {
+data class DirectedWeightedEdge(val begin: Int, val end: Int, val weight: Double) : WeightedEdge {
     override fun toString(): String {
         return "($begin -> $end w: $weight)"
     }
